fix infinite loop when python tests throw exceptions
In the python rand tests, when an exception is thrown it is caught and the tests is marked to be repeated with another random input.  If there is something fundamentally wrong with the test (like the test binary has an error) then this will cause it to go into an infinite loop.  Rather than trying to repeat the test with another input, this change causes it to just fail and move on.